Refactor SVONavigationData into modular implementation
Split SVONavigationData.cpp into multiple implementation files for pathfinding, drawing, generation, and details. Introduced FSVONavigationPathFinderImpl for pathfinding logic and moved debug info struct to its own file. Updated includes and initialization to use the new modular structure, improving maintainability and separation of concerns.
