+++ Source/SVONavigation/Private/SVONavigationData.cpp
@@ -1,40 +1,12 @@
 #include "SVONavigationData.h"
 
 #include "PathFinding/SVONavigationQueryFilterImpl.h"
-#include "PathFinding/SVOPathFinder.h"
-#include "PathFinding/SVOPathFindingAlgorithm.h"
-#include "SVOBoundsVolume.h"
-#include "SVONavDataRenderingComponent.h"
+#include "PathFinding/SVONavigationPathFinderImpl.h"
 #include "SVONavigationDataChunk.h"
-#include "SVONavigationDataGenerator.h"
-#include "SVONavigationSettings.h"
 #include "SVOVersion.h"
 
-#include <AI/NavDataGenerator.h>
-#include <DrawDebugHelpers.h>
-#include <EngineUtils.h>
-#include <NavMesh/NavMeshPath.h>
-#include <NavigationSystem.h>
-
-#if WITH_EDITOR
-#include <ObjectEditorUtils.h>
-#endif
-
-FSVOVolumeNavigationDataDebugInfos::FSVOVolumeNavigationDataDebugInfos() :
-    bDebugDrawBounds( false ),
-    bDebugDrawNodeCoords( false ),
-    bDebugDrawMortonCoords( false ),
-    bDebugDrawNodeAddresses( false ),
-    bDebugDrawNodeLocation( false ),
-    bDebugDrawLayers( false ),
-    LayerIndexToDraw( 0 ),
-    bDebugDrawSubNodes( false ),
-    bDebugDrawOccludedVoxels( true ),
-    bDebugDrawFreeVoxels( false ),
-    bDebugDrawNeighborLinks( false ),
-    bDebugDrawActivePaths( false )
-{
-}
+#include "AI/NavDataGenerator.h"
+#include "NavigationSystem.h"
 
 ASVONavigationData::ASVONavigationData() :
     Version( ESVOVersion::Latest )
@@ -43,20 +15,7 @@ ASVONavigationData::ASVONavigationData() :
 
     if ( !HasAnyFlags( RF_ClassDefaultObject ) )
     {
-        FindPathImplementation = FindPath;
-        /*FindHierarchicalPathImplementation = FindPath;
-
-        TestPathImplementation = TestPath;
-        TestHierarchicalPathImplementation = TestHierarchicalPath;*/
-
-        // RaycastImplementation = NavMeshRaycast;
-
-        // RecastNavMeshImpl = new FPImplRecastNavMesh( this );
-
-        //// add predefined areas up front
-        // SupportedAreas.Add( FSupportedAreaData( UNavArea_Null::StaticClass(), RECAST_NULL_AREA ) );
-        // SupportedAreas.Add( FSupportedAreaData( UNavArea_LowHeight::StaticClass(), RECAST_LOW_AREA ) );
-        // SupportedAreas.Add( FSupportedAreaData( UNavArea_Default::StaticClass(), RECAST_DEFAULT_AREA ) );
+        FindPathImplementation = FSVONavigationPathFinderImpl::FindPath;
     }
 }
 
@@ -160,593 +119,6 @@ void ASVONavigationData::CleanUp()
     ResetGenerator();
 }
 
-bool ASVONavigationData::NeedsRebuild() const
-{
-    const auto needs_rebuild = VolumeNavigationData.FindByPredicate( []( const FSVOVolumeNavigationData & data ) {
-        return !data.GetData().IsValid();
-    } ) != nullptr;
-
-    if ( NavDataGenerator.IsValid() )
-    {
-        return needs_rebuild || NavDataGenerator->GetNumRemaningBuildTasks() > 0;
-    }
-
-    return needs_rebuild;
-}
-
-void ASVONavigationData::EnsureBuildCompletion()
-{
-    Super::EnsureBuildCompletion();
-
-    // Doing this as a safety net solution due to UE-20646, which was basically a result of random
-    // over-releasing of default filter's shared pointer (it seemed). We might have time to get
-    // back to this time some time in next 3 years :D
-    RecreateDefaultFilter();
-}
-
-bool ASVONavigationData::SupportsRuntimeGeneration() const
-{
-    return (RuntimeGeneration == ERuntimeGenerationType::Dynamic);
-}
-
-bool ASVONavigationData::SupportsStreaming() const
-{
-    return ( RuntimeGeneration != ERuntimeGenerationType::Dynamic );
-}
-
-FNavLocation ASVONavigationData::GetRandomPoint( FSharedConstNavQueryFilter /*filter*/, const UObject * /*querier*/ ) const
-{
-    FNavLocation result;
-
-    const auto navigation_bounds_num = VolumeNavigationData.Num();
-
-    if ( navigation_bounds_num == 0 )
-    {
-        return result;
-    }
-
-    TArray< int > navigation_bounds_indices;
-    navigation_bounds_indices.Reserve( VolumeNavigationData.Num() );
-
-    for ( auto index = 0; index < navigation_bounds_num; index++ )
-    {
-        navigation_bounds_indices.Add( index );
-    }
-
-    // Shuffle the array
-    for ( int index = navigation_bounds_indices.Num() - 1; index > 0; --index )
-    {
-        const auto new_index = FMath::RandRange( 0, index );
-        Swap( navigation_bounds_indices[ index ], navigation_bounds_indices[ new_index ] );
-    }
-
-    do
-    {
-        const auto index = navigation_bounds_indices.Pop( EAllowShrinking::No );
-        const auto & volume_navigation_data = VolumeNavigationData[ index ];
-
-        const auto random_point = volume_navigation_data.GetRandomPoint();
-        if ( random_point.IsSet() )
-        {
-            result = random_point.GetValue();
-            break;
-        }
-    } while ( navigation_bounds_indices.Num() > 0 );
-
-    return result;
-}
-
-bool ASVONavigationData::GetRandomReachablePointInRadius( const FVector & origin, float radius, FNavLocation & out_result, FSharedConstNavQueryFilter filter, const UObject * querier ) const
-{
-    // :TODO:
-    ensure( false );
-    return false;
-}
-
-bool ASVONavigationData::GetRandomPointInNavigableRadius( const FVector & origin, float Radius, FNavLocation & out_result, FSharedConstNavQueryFilter filter, const UObject * querier ) const
-{
-    // :TODO:
-    ensure( false );
-    return false;
-}
-
-void ASVONavigationData::BatchRaycast( TArray< FNavigationRaycastWork > & workload, FSharedConstNavQueryFilter filter, const UObject * querier ) const
-{
-    // :TODO:
-    ensure( false );
-}
-
-bool ASVONavigationData::FindMoveAlongSurface( const FNavLocation & start_location, const FVector & target_position, FNavLocation & out_location, FSharedConstNavQueryFilter filter, const UObject * querier ) const
-{
-    // :TODO:
-    ensure( false );
-    return false;
-}
-
-bool ASVONavigationData::ProjectPoint( const FVector & point, FNavLocation & out_location, const FVector & extent, FSharedConstNavQueryFilter filter, const UObject * querier ) const
-{
-    if (VolumeNavigationData.IsEmpty())
-    {
-        return false;
-    }
-
-    // 1. Find the correct volume to search in.
-    const FSVOVolumeNavigationData* VolumeToSearch = nullptr;
-    for (const auto& Volume : VolumeNavigationData)
-    {
-        if (Volume.GetData().GetNavigationBounds().IsInside(point))
-        {
-            VolumeToSearch = &Volume;
-            break;
-        }
-    }
-
-    // Fallback: If the point is outside all volumes, find the closest volume to search within.
-    if (!VolumeToSearch)
-    {
-        float MinDistSq = -1.0f;
-        for (const auto& Volume : VolumeNavigationData)
-        {
-            const FBox& BoundingBox = Volume.GetData().GetNavigationBounds();
-            if (!BoundingBox.IsValid)
-                continue;
-
-            const float DistSq = BoundingBox.ComputeSquaredDistanceToPoint(point);
-            if (MinDistSq < 0 || DistSq < MinDistSq)
-            {
-                MinDistSq = DistSq;
-                VolumeToSearch = &Volume;
-            }
-        }
-    }
-
-    if ( !VolumeToSearch )
-    {
-        return false;
-    }
-
-    // 2. Clamp the search point to be within the volume's bounds.
-    FVector StartPoint = point;
-    const FBox& VolumeBounds = VolumeToSearch->GetData().GetNavigationBounds();
-    if (!VolumeBounds.IsInside(StartPoint))
-    {
-        StartPoint = VolumeBounds.GetClosestPointTo(StartPoint);
-    }
-
-    // 3. Attempt to find the initial node and check if it's already navigable.
-    FSVONodeAddress InitialAddress;
-    const bool bInitialNodeFound = VolumeToSearch->GetNodeAddressFromPosition(InitialAddress, StartPoint);
-    if (bInitialNodeFound && VolumeToSearch->IsNodeAddressNavigable(InitialAddress))
-    {
-        out_location.Location = StartPoint;
-        out_location.NodeRef = InitialAddress.GetNavNodeRef();
-        return true;
-    }
-
-    // 4. BFS Initialization: The start point is either in an occluded node or couldn't be resolved.
-    //    We must perform a search for the nearest navigable one.
-    TQueue<FSVONodeAddress> OpenList;
-    TSet<FSVONodeAddress> VisitedList;
-    const FBox SearchBounds = FBox::BuildAABB(StartPoint, extent);
-    
-    if (bInitialNodeFound)
-    {
-        // Start point is in an occluded node, begin search from there.
-        OpenList.Enqueue(InitialAddress);
-        VisitedList.Add(InitialAddress);
-    }
-    else
-    {
-        // The start point could not be resolved to any node.
-        // Seed the search with the nearest nodes to the start point instead of failing.
-        
-        TArray<FSVONodeAddress> SeedNodes;
-        // A small radius, just enough to find the immediate surrounding nodes.
-        const float SeedRadius = VolumeToSearch->GetData().GetLeafNodes().GetLeafNodeExtent() * 1.5f;
-        VolumeToSearch->FindNodesInSphere(StartPoint, SeedRadius, SeedNodes);
-        
-        if (SeedNodes.IsEmpty())
-        {
-             return false;
-        }
-
-        for (const FSVONodeAddress& SeedNode : SeedNodes)
-        {
-            if (!VisitedList.Contains(SeedNode))
-            {
-                 OpenList.Enqueue(SeedNode);
-                 VisitedList.Add(SeedNode);
-            }
-        }
-    }
-    
-    // 5. BFS Loop
-    FSVONodeAddress CurrentAddress;
-    while ( OpenList.Dequeue(CurrentAddress))
-    {
-        TArray<FSVONodeAddress> Neighbors;
-        VolumeToSearch->GetNodeNeighbors(Neighbors, CurrentAddress);
-
-        for (const FSVONodeAddress& NeighborAddress : Neighbors)
-        {
-            if (!VisitedList.Contains( NeighborAddress))
-            {
-                VisitedList.Add(NeighborAddress);
-
-                const FVector NeighborLocation = VolumeToSearch->GetNodePositionFromAddress(NeighborAddress, true);
-
-                if (!SearchBounds.IsInside(NeighborLocation))
-                {
-                    continue;
-                }
-
-                if (VolumeToSearch->IsNodeAddressNavigable(NeighborAddress))
-                {
-                    out_location.Location = NeighborLocation;
-                    out_location.NodeRef = NeighborAddress.GetNavNodeRef();
-                    return true;
-                }
-
-                OpenList.Enqueue(NeighborAddress);
-            }
-        }
-    }
-
-    // 6. Failure
-    return false;
-}
-
-void ASVONavigationData::BatchProjectPoints( TArray< FNavigationProjectionWork > & Workload, const FVector & Extent, FSharedConstNavQueryFilter Filter, const UObject * Querier ) const
-{
-    for (FNavigationProjectionWork& WorkItem : Workload)
-    {
-        WorkItem.bResult = ProjectPoint(WorkItem.Point, WorkItem.OutLocation, Extent, Filter, Querier);
-    }
-}
-
-void ASVONavigationData::BatchProjectPoints( TArray< FNavigationProjectionWork > & Workload, FSharedConstNavQueryFilter Filter, const UObject * Querier ) const
-{
-    for (FNavigationProjectionWork& WorkItem : Workload)
-    {
-        WorkItem.bResult = ProjectPoint(WorkItem.Point, WorkItem.OutLocation, FVector::ZeroVector, Filter, Querier);
-    }
-}
-
-ENavigationQueryResult::Type ASVONavigationData::CalcPathCost( const FVector & path_start, const FVector & path_end, FVector::FReal & out_path_cost, const FSharedConstNavQueryFilter filter, const UObject * querier ) const
-{
-    FVector::FReal path_length = 0.f;
-    return CalcPathLengthAndCost( path_start, path_end, path_length, out_path_cost, filter, querier );
-}
-
-ENavigationQueryResult::Type ASVONavigationData::CalcPathLength( const FVector & path_start, const FVector & path_end, FVector::FReal & out_path_length, const FSharedConstNavQueryFilter filter, const UObject * querier ) const
-{
-    FVector::FReal path_cost = 0.f;
-    return CalcPathLengthAndCost( path_start, path_end, out_path_length, path_cost, filter, querier );
-}
-
-ENavigationQueryResult::Type ASVONavigationData::CalcPathLengthAndCost( const FVector & path_start, const FVector & path_end, FVector::FReal & out_path_length, FVector::FReal & out_path_cost, FSharedConstNavQueryFilter filter, const UObject * querier ) const
-{
-    ENavigationQueryResult::Type result = ENavigationQueryResult::Invalid;
-
-    if ( ( path_start - path_end ).IsNearlyZero() )
-    {
-        out_path_length = 0.f;
-        return ENavigationQueryResult::Success;
-    }
-
-    auto * volume_navigation_data = GetVolumeNavigationDataContainingPoints( { path_start, path_end } );
-
-    if ( volume_navigation_data == nullptr )
-    {
-        return ENavigationQueryResult::Error;
-    }
-
-    const TSharedRef< FSVONavigationPath > navigation_path = MakeShareable( new FSVONavigationPath() );
-
-    result = FSVOPathFinder::GetPath( navigation_path.Get(), *this, path_start, path_end, filter );
-
-    if ( result == ENavigationQueryResult::Success || ( result == ENavigationQueryResult::Fail && navigation_path->IsPartial() ) )
-    {
-        out_path_length = navigation_path->GetLength();
-        out_path_cost = navigation_path->GetCost();
-    }
-
-    return result;
-}
-
-bool ASVONavigationData::DoesNodeContainLocation( NavNodeRef node_ref, const FVector & world_space_location ) const
-{
-    const FSVONodeAddress Address(node_ref);
-    if (!Address.IsValid())
-    {
-        return false;
-    }
-
-    for (const auto& Volume : VolumeNavigationData)
-    {
-        // A simple check to see if the location is even in this volume. This isn't perfect
-        // as a node from one volume could technically contain a point just inside another,
-        // but it's a reasonable optimization.
-        if (Volume.GetData().GetNavigationBounds().IsInside(world_space_location))
-        {
-            const FVector NodeLocation = Volume.GetNodePositionFromAddress(Address, true);
-            const float NodeExtent = Volume.GetNodeExtentFromNodeAddress(Address);
-            const FBox NodeBounds = FBox::BuildAABB(NodeLocation, FVector(NodeExtent));
-
-            if (NodeBounds.IsInsideOrOn(world_space_location))
-            {
-                return true;
-            }
-        }
-    }
-
-    return false;
-}
-
-UPrimitiveComponent * ASVONavigationData::ConstructRenderingComponent()
-{
-    return NewObject< USVONavDataRenderingComponent >( this, TEXT( "SVONavRenderingComp" ), RF_Transient );
-}
-
-void ASVONavigationData::OnStreamingLevelAdded( ULevel * level, UWorld * /*world*/ )
-{
-    // QUICK_SCOPE_CYCLE_COUNTER( STAT_RecastNavMesh_OnStreamingLevelAdded );
-
-    if ( SupportsStreaming() )
-    {
-        if ( USVONavigationDataChunk * navigation_data_chunk = GetNavigationDataChunk( level ) )
-        {
-            for ( const auto & chunk_nav_data : navigation_data_chunk->NavigationData )
-            {
-                if ( VolumeNavigationData.FindByPredicate( [ &chunk_nav_data ]( const auto & navigation_data ) {
-                         return chunk_nav_data.GetVolumeBounds() == navigation_data.GetVolumeBounds();
-                     } ) == nullptr )
-                {
-                    VolumeNavigationData.Add( chunk_nav_data );
-                }
-            }
-
-            RequestDrawingUpdate();
-        }
-    }
-}
-
-void ASVONavigationData::OnStreamingLevelRemoved( ULevel * level, UWorld * /*world*/ )
-{
-    // QUICK_SCOPE_CYCLE_COUNTER( STAT_RecastNavMesh_OnStreamingLevelRemoved );
-
-    if ( SupportsStreaming() )
-    {
-        if ( USVONavigationDataChunk * navigation_data_chunk = GetNavigationDataChunk( level ) )
-        {
-            for ( const auto & chunk_nav_data : navigation_data_chunk->NavigationData )
-            {
-                VolumeNavigationData.RemoveAllSwap( [ &chunk_nav_data ]( const auto & nav_data ) {
-                    return chunk_nav_data.GetVolumeBounds() == nav_data.GetVolumeBounds();
-                } );
-            }
-
-            RequestDrawingUpdate();
-        }
-    }
-}
-
-void ASVONavigationData::OnNavAreaChanged()
-{
-    Super::OnNavAreaChanged();
-}
-
-void ASVONavigationData::OnNavAreaAdded( const UClass * nav_area_class, int32 agent_index )
-{
-    Super::OnNavAreaAdded( nav_area_class, agent_index );
-}
-
-int32 ASVONavigationData::GetNewAreaID( const UClass * nav_area_class ) const
-{
-    return Super::GetNewAreaID( nav_area_class );
-}
-
-int32 ASVONavigationData::GetMaxSupportedAreas() const
-{
-    return 32;
-}
-
-bool ASVONavigationData::IsNodeRefValid( const NavNodeRef node_ref ) const
-{
-    return FSVONodeAddress( node_ref ).IsValid();
-}
-
-void ASVONavigationData::TickActor( const float delta_time, const ELevelTick tick, FActorTickFunction & this_tick_function )
-{
-    Super::TickActor( delta_time, tick, this_tick_function );
-
-#if ENABLE_DRAW_DEBUG
-
-    if ( bEnableDrawing && DebugInfos.bDebugDrawActivePaths )
-    {
-        for ( auto active_path : ActivePaths )
-        {
-            if ( !active_path.IsValid() )
-            {
-                continue;
-            }
-
-            const TSharedPtr< FNavigationPath, ESPMode::ThreadSafe > active_path_ptr = active_path.Pin();
-            const auto & path_points = active_path_ptr->GetPathPoints();
-
-            for ( auto path_point_index = 1; path_point_index < path_points.Num(); ++path_point_index )
-            {
-                const auto & from = path_points[ path_point_index - 1 ].Location;
-                const auto & to = path_points[ path_point_index ].Location;
-
-                DrawDebugLine( GetWorld(), from, to, FColor::Red, false, -1, SDPG_World, 5.0f );
-                DrawDebugCone( GetWorld(), to, from - to, 50.0f, 0.25f, 0.25f, 16, FColor::Red, false, -1, SDPG_World, 5.0f );
-            }
-        }
-    }
-
-#endif
-}
-
-#if WITH_EDITOR
-void ASVONavigationData::PostEditChangeProperty( FPropertyChangedEvent & property_changed_event )
-{
-    Super::PostEditChangeProperty( property_changed_event );
-
-    if ( property_changed_event.Property == nullptr )
-    {
-        return;
-    }
-
-    if ( property_changed_event.Property != nullptr )
-    {
-        const FName category_name = FObjectEditorUtils::GetCategoryFName( property_changed_event.Property );
-        static const FName NAME_Generation = FName( TEXT( "Generation" ) );
-        static const FName NAME_Query = FName( TEXT( "Query" ) );
-
-        if ( category_name == NAME_Generation )
-        {
-            if ( auto * settings = GetDefault< USVONavigationSettings >() )
-            {
-                if ( !HasAnyFlags( RF_ClassDefaultObject ) && settings->bNavigationAutoUpdateEnabled )
-                {
-                    RebuildAll();
-                }
-            }
-        }
-        else if ( category_name == NAME_Query )
-        {
-            RecreateDefaultFilter();
-        }
-    }
-}
-
-bool ASVONavigationData::ShouldExport()
-{
-    return false;
-}
-#endif
-
-#if !UE_BUILD_SHIPPING
-uint32 ASVONavigationData::LogMemUsed() const
-{
-    const auto super_mem_used = Super::LogMemUsed();
-
-    auto navigation_mem_size = 0;
-    for ( const auto & nav_bounds_data : VolumeNavigationData )
-    {
-        const auto octree_data_mem_size = nav_bounds_data.GetData().GetAllocatedSize();
-        navigation_mem_size += octree_data_mem_size;
-    }
-    const auto mem_used = super_mem_used + navigation_mem_size;
-
-    UE_LOG( LogNavigation, Warning, TEXT( "%s: ASVONavigationData: %u\n    self: %d" ), *GetName(), mem_used, sizeof( ASVONavigationData ) );
-
-    return mem_used;
-}
-#endif
-
-void ASVONavigationData::ConditionalConstructGenerator()
-{
-    ResetGenerator();
-
-    UWorld * world = GetWorld();
-    check( world );
-    const bool requires_generator = SupportsRuntimeGeneration() || !world->IsGameWorld();
-
-    if ( !requires_generator )
-    {
-        return;
-    }
-
-    if ( FSVONavigationDataGenerator * generator = new FSVONavigationDataGenerator( *this ) )
-    {
-        NavDataGenerator = MakeShareable( static_cast< FNavDataGenerator * >( generator ) );
-        generator->Init();
-    }
-}
-
-void ASVONavigationData::RequestDrawingUpdate( const bool force )
-{
-#if !UE_BUILD_SHIPPING
-    if ( force || USVONavDataRenderingComponent::IsNavigationShowFlagSet( GetWorld() ) )
-    {
-        if ( force )
-        {
-            if ( USVONavDataRenderingComponent * rendering_component = Cast< USVONavDataRenderingComponent >( RenderingComp ) )
-            {
-                rendering_component->ForceUpdate();
-            }
-        }
-
-        DECLARE_CYCLE_STAT( TEXT( "FSimpleDelegateGraphTask.Requesting SVO navmesh redraw" ),
-            STAT_FSimpleDelegateGraphTask_RequestingNavmeshRedraw,
-            STATGROUP_TaskGraphTasks );
-
-        FSimpleDelegateGraphTask::CreateAndDispatchWhenReady(
-            FSimpleDelegateGraphTask::FDelegate::CreateUObject( this, &ASVONavigationData::UpdateDrawing ),
-            GET_STATID( STAT_FSimpleDelegateGraphTask_RequestingNavmeshRedraw ),
-            nullptr,
-            ENamedThreads::GameThread );
-    }
-#endif // !UE_BUILD_SHIPPING
-}
-
-FBox ASVONavigationData::GetBoundingBox() const
-{
-    FBox bounding_box( ForceInit );
-
-    for ( const auto & bounds : VolumeNavigationData )
-    {
-        bounding_box += bounds.GetData().GetNavigationBounds();
-    }
-
-    return bounding_box;
-}
-
-void ASVONavigationData::RemoveDataInBounds( const FBox & bounds )
-{
-    VolumeNavigationData.RemoveAllSwap( [ &bounds ]( const FSVOVolumeNavigationData & data ) {
-        return data.GetVolumeBounds() == bounds;
-    } );
-}
-
-void ASVONavigationData::AddVolumeNavigationData( FSVOVolumeNavigationData data )
-{
-    for ( TActorIterator< ASVOBoundsVolume > iterator( GetWorld(), ASVOBoundsVolume::StaticClass() ); iterator; ++iterator )
-    {
-        const auto * volume = *iterator;
-
-        if ( volume->GetComponentsBoundingBox( true ) == data.GetVolumeBounds() )
-        {
-            data.SetVolumeNavigationQueryFilter( volume->GetVolumeNavigationQueryFilter() );
-            break;
-        }
-    }
-
-    VolumeNavigationData.Emplace( MoveTemp( data ) );
-}
-
-const FSVOVolumeNavigationData * ASVONavigationData::GetVolumeNavigationDataContainingPoints( const TArray< FVector > & points ) const
-{
-    return VolumeNavigationData.FindByPredicate( [ this, &points ]( const FSVOVolumeNavigationData & data ) {
-        const auto & bounds = data.GetData().GetNavigationBounds();
-        for ( const auto & point : points )
-        {
-            if ( !bounds.IsInside( point ) )
-            {
-                return false;
-            }
-        }
-        return true;
-    } );
-}
-
-void ASVONavigationData::UpdateNavVersion()
-{
-    Version = ESVOVersion::Latest;
-}
-
 void ASVONavigationData::SerializeSVOData( FArchive & archive, ESVOVersion version )
 {
     if ( archive.IsLoading() )
@@ -808,300 +180,7 @@ void ASVONavigationData::SerializeSVOData( FArchive & archive, ESVOVersion versi
     }
 }
 
-void ASVONavigationData::CheckToDiscardSubLevelNavData( const UNavigationSystemBase & navigation_system )
-{
-    if ( const auto * world = GetWorld() )
-    {
-        if ( const auto * nav_sys = Cast< UNavigationSystemV1 >( &navigation_system ) )
-        {
-            // Get rid of instances saved within levels that are streamed-in
-            if ( GEngine->IsSettingUpPlayWorld() == false // this is a @HACK
-                 && ( world->PersistentLevel != GetLevel() )
-                 // If we are cooking, then let them all pass.
-                 // They will be handled at load-time when running.
-                 && ( IsRunningCommandlet() == false ) )
-            {
-                UE_LOG( LogNavigation, Verbose, TEXT( "%s Discarding %s due to it not being part of PersistentLevel." ), ANSI_TO_TCHAR( __FUNCTION__ ), *GetFullNameSafe( this ) );
-
-                // Marking self for deletion
-                CleanUpAndMarkPendingKill();
-            }
-        }
-    }
-}
-
 void ASVONavigationData::RecreateDefaultFilter() const
 {
     DefaultQueryFilter->SetFilterType< FSVONavigationQueryFilterImpl >();
-}
-
-void ASVONavigationData::UpdateDrawing() const
-{
-#if !UE_BUILD_SHIPPING
-    if ( USVONavDataRenderingComponent * rendering_component = Cast< USVONavDataRenderingComponent >( RenderingComp ) )
-    {
-        if ( rendering_component->GetVisibleFlag() && ( rendering_component->UpdateIsForced() || USVONavDataRenderingComponent::IsNavigationShowFlagSet( GetWorld() ) ) )
-        {
-            rendering_component->MarkRenderStateDirty();
-        }
-    }
-#endif /
-}
-
-void ASVONavigationData::ResetGenerator( const bool cancel_build )
-{
-    if ( NavDataGenerator.IsValid() )
-    {
-        if ( cancel_build )
-        {
-            NavDataGenerator->CancelBuild();
-        }
-
-        NavDataGenerator.Reset();
-    }
-}
-
-void ASVONavigationData::OnNavigationDataUpdatedInBounds( const TArray< FBox > & updated_bounds )
-{
-    InvalidateAffectedPaths( updated_bounds );
-}
-
-void ASVONavigationData::ClearNavigationData()
-{
-    VolumeNavigationData.Reset();
-    RequestDrawingUpdate();
-}
-
-void ASVONavigationData::BuildNavigationData()
-{
-    RebuildAll();
-}
-
-void ASVONavigationData::InvalidateAffectedPaths( const TArray< FBox > & updated_bounds )
-{
-    const int32 paths_count = ActivePaths.Num();
-    const int32 updated_bounds_count = updated_bounds.Num();
-
-    if ( updated_bounds_count == 0 || paths_count == 0 )
-    {
-        return;
-    }
-
-    // Paths can be registered from async pathfinding thread.
-    // Theoretically paths are invalidated synchronously by the navigation system
-    // before starting async queries task but protecting ActivePaths will make
-    // the system safer in case of future timing changes.
-    {
-        FScopeLock path_lock( &ActivePathsLock );
-
-        FNavPathWeakPtr * weak_path_ptr = ( ActivePaths.GetData() + paths_count - 1 );
-
-        for ( int32 path_index = paths_count - 1; path_index >= 0; --path_index, --weak_path_ptr )
-        {
-            FNavPathSharedPtr shared_path = weak_path_ptr->Pin();
-            if ( !weak_path_ptr->IsValid() )
-            {
-                ActivePaths.RemoveAtSwap( path_index, 1, EAllowShrinking::No );
-            }
-            else
-            {
-                const FNavigationPath * path = shared_path.Get();
-                if ( !path->IsReady() || path->GetIgnoreInvalidation() )
-                {
-                    // path not filled yet or doesn't care about invalidation
-                    continue;
-                }
-
-                for ( const auto & path_point : path->GetPathPoints() )
-                {
-                    if ( updated_bounds.FindByPredicate( [ &path_point ]( const FBox & bounds ) {
-                             return bounds.IsInside( path_point.Location );
-                         } ) != nullptr )
-                    {
-                        shared_path->Invalidate();
-                        ActivePaths.RemoveAtSwap( path_index, 1, EAllowShrinking::No );
-
-                        break;
-                    }
-                }
-
-                if ( !shared_path->IsValid() )
-                {
-                    break;
-                }
-            }
-        }
-    }
-}
-
-void ASVONavigationData::OnNavigationDataGenerationFinished()
-{
-    if ( UWorld * world = GetWorld() )
-    {
-        if ( IsValid( world ) )
-        {
-#if WITH_EDITOR
-            // For navmeshes that support streaming create navigation data holders in each streaming level
-            // so parts of navmesh can be streamed in/out with those levels
-            if ( !world->IsGameWorld() )
-            {
-                const auto & levels = world->GetLevels();
-
-                for ( auto * level : levels )
-                {
-                    if ( level->IsPersistentLevel() )
-                    {
-                        continue;
-                    }
-
-                    USVONavigationDataChunk * navigation_data_chunk = GetNavigationDataChunk( level );
-
-                    if ( SupportsStreaming() )
-                    {
-                        // We use navigation volumes that belongs to this streaming level to find tiles we want to save
-                        const auto & level_nav_bounds = GetNavigableBoundsInLevel( level );
-
-                        TArray< int32 > navigation_data_indices;
-                        navigation_data_indices.Reserve( level_nav_bounds.Num() );
-
-                        for ( const auto & nav_bounds : level_nav_bounds )
-                        {
-                            const auto index = VolumeNavigationData.IndexOfByPredicate( [ &nav_bounds ]( const FSVOVolumeNavigationData & data ) {
-                                const auto & bounds = data.GetData().GetVolumeBounds();
-                                return bounds == nav_bounds;
-                            } );
-
-                            if ( index != INDEX_NONE )
-                            {
-                                navigation_data_indices.Add( index );
-                            }
-                        }
-
-                        if ( navigation_data_indices.Num() > 0 )
-                        {
-                            // Create new chunk only if we have something to save in it
-                            if ( navigation_data_chunk == nullptr )
-                            {
-                                navigation_data_chunk = NewObject< USVONavigationDataChunk >( level );
-                                navigation_data_chunk->NavigationDataName = GetFName();
-                                level->NavDataChunks.Add( navigation_data_chunk );
-                            }
-
-                            for ( const auto index : navigation_data_indices )
-                            {
-                                navigation_data_chunk->AddNavigationData( VolumeNavigationData[ index ] );
-                            }
-
-                            navigation_data_chunk->MarkPackageDirty();
-                            continue;
-                        }
-                    }
-
-                    // It's hack. That check should not be there.
-                    // When calling FNavigationSystem::Build, all streaming levels should be loaded and visible for the navigation to be built. That's how it works for ReCast
-                    // But since svo nav data always resolves to a box bigger than the nav bounds volume, it's possible that when building navigation for a volume in a streaming
-                    // level, the box would encompasses geometry of another level which should not be visible.
-                    // The solution we use in our game is to use a BuildIncremental function on a custom navigation system, which never calls FNavigationSystem::DiscardNavigationDataChunks
-                    // In a commandlet we load streaming levels by batch, build navigation for those levels only, then load another batch of levels, build navigation for those levels, etc...
-                    // This means that this function ASVONavigationData::OnNavigationDataGenerationFinished is called after navigation is built for each batch of levels
-                    // and that also means that after the last batch of levels is processed, we would release the navigation data for each previous batch of levels
-                    if ( !IsRunningCommandlet() )
-                    {
-                        // stale data that is left in the level
-                        if ( navigation_data_chunk != nullptr )
-                        {
-                            navigation_data_chunk->ReleaseNavigationData();
-                            navigation_data_chunk->MarkPackageDirty();
-                            level->NavDataChunks.Remove( navigation_data_chunk );
-                        }
-                    }
-                }
-            }
-
-            // force navmesh drawing update
-            RequestDrawingUpdate( /*bForce=*/true );
-#endif // WITH_EDITOR
-
-            UNavigationSystemV1 * NavSys = FNavigationSystem::GetCurrent< UNavigationSystemV1 >( world );
-            if ( NavSys )
-            {
-                NavSys->OnNavigationGenerationFinished( *this );
-            }
-
-            DataInfos.Infos.Reset();
-
-            for ( const auto & bounds_navigation_data : VolumeNavigationData )
-            {
-                auto & navigation_data_infos = DataInfos.Infos.AddDefaulted_GetRef();
-                navigation_data_infos.VolumeLocation = bounds_navigation_data.GetVolumeBounds().GetCenter();
-                navigation_data_infos.LayerCount = bounds_navigation_data.GetData().GetLayerCount();
-                navigation_data_infos.bHasNavigationData = bounds_navigation_data.GetData().IsValid();
-            }
-        }
-    }
-}
-
-USVONavigationDataChunk * ASVONavigationData::GetNavigationDataChunk( ULevel * level ) const
-{
-    const auto this_name = GetFName();
-
-    if ( const auto * result = level->NavDataChunks.FindByPredicate( [ & ]( const UNavigationDataChunk * chunk ) {
-             return chunk->NavigationDataName == this_name;
-         } ) )
-    {
-        return Cast< USVONavigationDataChunk >( *result );
-    }
-
-    return nullptr;
-}
-
-FPathFindingResult ASVONavigationData::FindPath( const FNavAgentProperties & /*agent_properties*/, const FPathFindingQuery & path_finding_query )
-{
-    const auto * self = Cast< ASVONavigationData >( path_finding_query.NavData.Get() );
-
-    if ( self == nullptr )
-    {
-        return ENavigationQueryResult::Error;
-    }
-
-    FPathFindingResult result( ENavigationQueryResult::Error );
-
-    FNavigationPath * navigation_path = path_finding_query.PathInstanceToFill.Get();
-    FSVONavigationPath * svo_navigation_path = navigation_path != nullptr
-                                                   ? navigation_path->CastPath< FSVONavigationPath >()
-                                                   : nullptr;
-
-    if ( svo_navigation_path != nullptr )
-    {
-        result.Path = path_finding_query.PathInstanceToFill;
-        svo_navigation_path->ResetForRepath();
-    }
-    else
-    {
-        result.Path = self->CreatePathInstance< FSVONavigationPath >( path_finding_query );
-        navigation_path = result.Path.Get();
-        svo_navigation_path = navigation_path != nullptr
-                                  ? navigation_path->CastPath< FSVONavigationPath >()
-                                  : nullptr;
-    }
-
-    if ( navigation_path != nullptr )
-    {
-        if ( path_finding_query.QueryFilter.IsValid() )
-        {
-            const FVector adjusted_end_location = path_finding_query.EndLocation; // navigation_filter->GetAdjustedEndLocation( path_finding_query.EndLocation );
-            if ( ( path_finding_query.StartLocation - adjusted_end_location ).IsNearlyZero() )
-            {
-                result.Path->GetPathPoints().Reset();
-                result.Path->GetPathPoints().Add( FNavPathPoint( adjusted_end_location ) );
-                result.Result = ENavigationQueryResult::Success;
-            }
-            else
-            {
-                result.Result = FSVOPathFinder::GetPath( *svo_navigation_path, *self, path_finding_query.StartLocation, adjusted_end_location, path_finding_query.QueryFilter );
-            }
-        }
-    }
-
-    return result;
 }
